Removed unneeded headers from MainWindow.cpp
Since the code in hbshaper.h then ends up in MainWindow.o and
MainView.o, the linker will complain that there are multible
clashing symbols and refuse to link. So hbshaper.h as is can only
be included from one source file at the moment. To fix this, it
would have to be devided into an .h with just the class declaration
and a .cpp with the implementation. Like the other files.